rename/#201 : 인스턴스 구조 리팩토링에따른 파일명을 수정한다
