+++ frontend/src/apis/dm/index.ts
@@ -1,5 +1,5 @@
 import { DMListResponseDto, DMResponseDto } from '@/apis/dm/dto';
-import { directMessageApi } from '@/lib/https';
+import { directMessageApi } from '@/lib/clients';
 
 export const getDMList = async (): Promise<DMListResponseDto> => {
   const { data } = await directMessageApi.get('');
